% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_class.R
\name{bi_class}
\alias{bi_class}
\title{Create Classes for Bivariate Maps}
\usage{
bi_class(.data, x, y)
}
\arguments{
\item{.data}{A data frame, tibble, or \code{sf} object}

\item{x}{The \code{x} variable}

\item{y}{The \code{y} variable}
}
\value{
A copy of \code{.data} with a new variable \code{bs_class} that contains
    combinations of values that correspond to an observations values for \code{x}
    and \code{y}. This is the basis for applying a bivariate color palette.
}
\description{
Creates mapping classes for a 3x3 bivariate map.
}
