% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_legend.R
\name{bi_legend}
\alias{bi_legend}
\title{Create Object for Drawing Legend}
\usage{
bi_legend(pal)
}
\arguments{
\item{pal}{A palette name}
}
\description{
Creates a tibble that can be used as the basis for
    drawing a legend.
}
