% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_pal.R
\name{bi_pal}
\alias{bi_pal}
\title{Print Hex Values for Palette}
\usage{
bi_pal(pal)
}
\arguments{
\item{pal}{The name of the palette to print hex values for.}
}
\description{
Prints the hex values for a specified palette.
}
