% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_theme.R
\name{bi_theme}
\alias{bi_theme}
\title{Basic Theme for Bivariate Mapping}
\usage{
bi_theme(...)
}
\arguments{
\item{...}{Arguments to pass on to \code{ggplot2}'s \code{theme} function}
}
\description{
A theme for creating a simple, clean bivariate map.
}
