% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_class.R
\name{bi_class}
\alias{bi_class}
\title{Create Classes for Bivariate Maps}
\usage{
bi_class(.data, x, y, style = "tercile", keep_factors = FALSE)
}
\arguments{
\item{.data}{A data frame, tibble, or \code{sf} object}

\item{x}{The \code{x} variable}

\item{y}{The \code{y} variable}

\item{style}{A string identifying the style used to calculate breaks. Currently
supported styles are \code{"tercile"} (default), \code{"equal"}, \code{"fisher"},
and \code{"jenks"}.}

\item{keep_factors}{A logical scalar; if \code{TRUE}, the fintermediate factor
variables created as part of the calculation of \code{bs_class} will be
retained. If \code{FALSE} (default), they will not be returned.}
}
\value{
A copy of \code{.data} with a new variable \code{bs_class} that contains
    combinations of values that correspond to an observations values for \code{x}
    and \code{y}. This is the basis for applying a bivariate color palette.
}
\description{
Creates mapping classes for a 3x3 bivariate map.
}
