% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_legend.R
\name{bi_legend}
\alias{bi_legend}
\title{Create Object for Drawing Legend}
\usage{
bi_legend(pal, xlab, ylab, size)
}
\arguments{
\item{pal}{A palette name}

\item{xlab}{The x axis label}

\item{ylab}{The y axis label}

\item{size}{Size of axis labels}
}
\description{
Creates a tibble that can be used as the basis for
    drawing a legend.
}
