% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_legend.R
\name{bi_legend}
\alias{bi_legend}
\title{Create Object for Drawing Legend}
\usage{
bi_legend(pal, dim = 3, xlab, ylab, size, flip_axes = FALSE,
    rotate_pal = FALSE, pad_width = NA, pad_color)
}
\arguments{
\item{pal}{A palette name or a vector containing a custom palette. See
the help file for \code{bi_pal} for complete list of built-in palette
names. If you are providing a custom palette, it must follow the formatting
described in the 'Advanced Options' vignette.}

\item{dim}{The dimensions of the palette. To use the built-in palettes,
    this value must be either \code{2}, \code{3}, or \code{4}. A value of
    \code{3}, for example, would be used to create a three-by-three bivariate
    map with a total of 9 classes.

    If you are using a custom palette, this value may be larger (though these
    maps can be very hard to interpret). See the 'Advanced Options' vignette
    for details on the relationship between \code{dim} values and palette size.}

\item{xlab}{Text for desired x axis label on legend}

\item{ylab}{Text for desired y axis label on legend}

\item{size}{A numeric scalar; size of axis labels}

\item{flip_axes}{A logical scalar; if \code{TRUE} the axes of the palette
will be flipped. If \code{FALSE} (default), the palette will be displayed
on its original axes. Custom palettes with 'dim' greater
than 4 cannot take advantage of flipping axes.}

\item{rotate_pal}{A logical scalar; if \code{TRUE} the palette will be
rotated 180 degrees. If \code{FALSE} (default), the palette will be
displayed in its original orientation. Custom palettes with 'dim' greater
than 4 cannot take advantage of palette rotation.}

\item{pad_width}{An optional numeric scalar; controls the width of padding
between tiles in the legend}

\item{pad_color}{An optional character scalar; controls the color of padding
between tiles in the legend}
}
\value{
A \code{ggplot} object with a bivariate legend.
}
\description{
Creates a \code{ggplot} object containing a legend that is
    specific to bivariate mapping.
}
\examples{
# sample 3x3 legend
legend <- bi_legend(pal = "GrPink",
                    dim = 3,
                    xlab = "Higher \% White ",
                    ylab = "Higher Income ",
                    size = 16)

# print legend
legend

}
\seealso{
bi_pal
}
