% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_pal.R
\name{bi_pal}
\alias{bi_pal}
\title{Add Palette Hex Values to Data}
\usage{
bi_pal(.data, pal)
}
\arguments{
\item{.data}{A \code{sf} object}

\item{pal}{The name of the palette to apply.}
}
\description{
Adding the palette hex values allows you to use \code{sale_fill_identity}
    to plot the bivariate data.
}
