% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_scale_color.R
\name{bi_scale_color}
\alias{bi_scale_color}
\title{Apply Bivariate Color to ggplot Object}
\usage{
bi_scale_color(pal, dim = 3, ...)
}
\arguments{
\item{pal}{Either palette name (one of \code{"Brown"}, \code{"DkBlue"},
\code{"DkCyan"}, \code{"DkViolet"}, or \code{"GrPink"}) or a custom palette
object created with \code{\link{bi_pal_manual}}.}

\item{dim}{The dimensions of the palette, either \code{2} for a two-by-two palette or
\code{3} for a three-by-three palette.}

\item{...}{Arguments to pass to \code{\link{scale_fill_manual}}}
}
\value{
A \code{ggplot} object with the given bivariate palette applied to the data.
}
\description{
Applies the selected palette as the color aesthetic when \code{\link{geom_sf}}
    is used and the \code{bi_class} variable is given as the \code{color} in the aesthetic
    mapping.
}
