% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_scale_fill.R
\name{bi_scale_fill}
\alias{bi_scale_fill}
\title{Apply Bivariate Scale to ggplot Object}
\usage{
bi_scale_fill(pal, ...)
}
\arguments{
\item{pal}{A palette name; one of \code{"Brown"}, \code{"DkBlue"},
\code{"DkCyan"}, \code{"DkViolet"}, or \code{"GrPink"}.}

\item{...}{Arguments to pass to \code{\link{scale_fill_manual}}}
}
\description{
Applies the selected palette as the fill aesthetic when \code{\link{geom_sf}}
    is used and the \code{bi_class} variable is used as the \code{fill} in the aesthetic
    mapping.
}
\examples{
# load dependency
library(ggplot2)

# add breaks
data <- bi_class(stl_race_income, x = pctWhite, y = medInc)

# create map
ggplot() +
  geom_sf(data = data, aes(fill = bi_class), color = "white", size = 0.1, show.legend = FALSE) +
  bi_scale_fill(pal = "GrPink")

}
